requirements.txt
alembic==1.18.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
feedparser==6.0.12
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.59.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
playwright==1.57.0
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
squarify==0.4.4
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vnai==2.3.0
vnstock==3.3.1
vnstock3==3.2.1
vnstock_ezchart==0.0.3
websockets==15.0.1
wordcloud==1.9.5
zipp==3.23.0
